RBQA-F003-AL1 rename message received emitter module
