+++ crates/atm-daemon/src/message_received_emitter.rs
